refactor(courtfinder): mark sensitive file server-only
